implemented support for creating folders and play-lists fixed a bug that occurred when a property for an object didn't exist cleaned up some parts of the code made the c++ overall more robust (more checking for invalid arguments)

+++ c++/src/Com.cpp
@@ -51,7 +51,7 @@ JNIEXPORT jobject JNICALL Java_be_derycke_pieter_com_COM_CoCreateInstance
 
 	//niet 100% want een dword is een unsigned long een een jlong een signed long
 	//maar doet het voor nu wel
-	hr = CoCreateInstance(rclsid, (LPUNKNOWN)pUnkOuter, dwClsContext, riid, &reference);
+	hr = CoCreateInstance(rclsid, (LPUNKNOWN)pUnkOuter, (DWORD)dwClsContext, riid, &reference);
 
 	if(SUCCEEDED(hr))
 	{
